+++ runtime/integration-tests/src/cases/precompile.rs
@@ -68,7 +68,7 @@ fn axelar_precompile_execute<T: Runtime>() {
 		)
 		.unwrap();
 
-		let msg = Message::Transfer {
+		let msg = Message::TransferAssets {
 			currency: general_currency_id,
 			sender: derived_sender_account.clone().into(),
 			receiver: derived_receiver_account.clone().into(),
